refactor(server): split schemas.py into a schemas/ package
The single 1.5k-line schemas.py held 112 Pydantic models across many
unrelated API domains. Move them into focused modules under a new
server/schemas/ package (repository, pages, graph, git, decisions,
workspace, ownership, c4, architecture, blast_radius, etc.), with the
package __init__ re-exporting every model so all existing
'from repowise.server.schemas import X' imports keep working unchanged.

Collapse the three byte-identical graph-node response models
(GraphNodeResponse / DeadCodeGraphNodeResponse / HotFilesNodeResponse)
via inheritance: the latter two now extend GraphNodeResponse and add
their single extra field, preserving field order and serialized output.

Add a characterization test covering the facade surface and pinning the
graph-node field order and JSON serialization.
