+++ packages/server/src/repowise/server/schemas/__init__.py
@@ -1,0 +1,270 @@
+"""Pydantic request/response models for the repowise REST API.
+
+This package was split out of the former single ``schemas.py`` module. Every
+public model is re-exported here so existing
+``from repowise.server.schemas import X`` imports keep working unchanged.
+"""
+
+from __future__ import annotations
+
+from .architecture import (
+    ArchEdgeResponse,
+    ArchitectureViewResponse,
+    ArchLayerResponse,
+    ArchNodeResponse,
+    ArchTourStepResponse,
+)
+from .blast_radius import (
+    BlastRadiusRequest,
+    BlastRadiusResponse,
+    CochangeWarning,
+    DirectRiskEntry,
+    ReviewerEntry,
+    TransitiveEntry,
+)
+from .c4 import (
+    C4ComponentResponse,
+    C4ContainerResponse,
+    C4ExternalSystemResponse,
+    C4L1Response,
+    C4L2Response,
+    C4L3Response,
+    C4PersonResponse,
+    C4RelationResponse,
+    C4SystemResponse,
+)
+from .chat import (
+    ChatMessageResponse,
+    ChatRequest,
+    ConversationResponse,
+)
+from .code_quality import (
+    DeadCodeFindingResponse,
+    DeadCodePatchRequest,
+    DeadCodeSummaryResponse,
+    RepoStatsResponse,
+    SecurityFindingResponse,
+)
+from .decisions import (
+    DecisionCodeEdge,
+    DecisionCreate,
+    DecisionEvidenceResponse,
+    DecisionGraphEdge,
+    DecisionGraphNode,
+    DecisionGraphResponse,
+    DecisionLineageEntry,
+    DecisionRecordResponse,
+    DecisionStatusUpdate,
+)
+from .git import (
+    GitMetadataResponse,
+    GitSummaryResponse,
+    HotspotResponse,
+    OwnershipEntry,
+)
+from .graph import (
+    ArchitectureEdgeResponse,
+    ArchitectureGraphResponse,
+    ArchitectureNodeResponse,
+    DeadCodeGraphNodeResponse,
+    DeadCodeGraphResponse,
+    EgoGraphResponse,
+    GraphEdgeResponse,
+    GraphExportResponse,
+    GraphNodeResponse,
+    HotFilesGraphResponse,
+    HotFilesNodeResponse,
+    ModuleEdgeResponse,
+    ModuleGraphResponse,
+    ModuleNodeResponse,
+    NodeSearchResult,
+)
+from .health import (
+    CoordinatorHealthResponse,
+    HealthResponse,
+)
+from .intelligence import (
+    CallerCalleeEntry,
+    CallersCalleesResponse,
+    CommunityDetailResponse,
+    CommunityMember,
+    CommunitySummaryItem,
+    ExecutionFlowEntry,
+    ExecutionFlowsResponse,
+    GraphMetricsResponse,
+    NeighboringCommunity,
+    SymbolNodeSummary,
+)
+from .ownership import (
+    KnowledgeMapOwner,
+    KnowledgeMapResponse,
+    KnowledgeMapSilo,
+    KnowledgeMapTarget,
+    ModuleHealthDetail,
+    ModuleHealthOwner,
+    ModuleHealthSummary,
+    OwnerCoAuthor,
+    OwnerFileEntry,
+    OwnerListEntry,
+    OwnerModuleRollup,
+    OwnerProfileResponse,
+    ReviewerSuggestion,
+    ReviewerSuggestionsResponse,
+)
+from .pages import (
+    JobResponse,
+    PageResponse,
+    PageVersionResponse,
+)
+from .pagination import Paginated
+from .repository import (
+    RepoCreate,
+    RepoResponse,
+    RepoUpdate,
+)
+from .search import (
+    SearchRequest,
+    SearchResultResponse,
+)
+from .symbols import (
+    SymbolImportanceComponents,
+    SymbolResponse,
+)
+from .ui_helpers import (
+    CostGroupResponse,
+    CostSummaryResponse,
+    SetActiveProviderRequest,
+    SetApiKeyRequest,
+    WebhookResponse,
+)
+from .workspace import (
+    WorkspaceCoChangeEntry,
+    WorkspaceCoChangesResponse,
+    WorkspaceContractEntry,
+    WorkspaceContractLinkEntry,
+    WorkspaceContractsResponse,
+    WorkspaceContractSummary,
+    WorkspaceCrossRepoSummary,
+    WorkspaceGraphEdge,
+    WorkspaceGraphNode,
+    WorkspaceGraphResponse,
+    WorkspaceRepoEntry,
+    WorkspaceResponse,
+    WorkspaceSyncResponse,
+    WorkspaceSyncResult,
+)
+
+__all__ = [
+    "ArchEdgeResponse",
+    "ArchLayerResponse",
+    "ArchNodeResponse",
+    "ArchTourStepResponse",
+    "ArchitectureEdgeResponse",
+    "ArchitectureGraphResponse",
+    "ArchitectureNodeResponse",
+    "ArchitectureViewResponse",
+    "BlastRadiusRequest",
+    "BlastRadiusResponse",
+    "C4ComponentResponse",
+    "C4ContainerResponse",
+    "C4ExternalSystemResponse",
+    "C4L1Response",
+    "C4L2Response",
+    "C4L3Response",
+    "C4PersonResponse",
+    "C4RelationResponse",
+    "C4SystemResponse",
+    "CallerCalleeEntry",
+    "CallersCalleesResponse",
+    "ChatMessageResponse",
+    "ChatRequest",
+    "CochangeWarning",
+    "CommunityDetailResponse",
+    "CommunityMember",
+    "CommunitySummaryItem",
+    "ConversationResponse",
+    "CoordinatorHealthResponse",
+    "CostGroupResponse",
+    "CostSummaryResponse",
+    "DeadCodeFindingResponse",
+    "DeadCodeGraphNodeResponse",
+    "DeadCodeGraphResponse",
+    "DeadCodePatchRequest",
+    "DeadCodeSummaryResponse",
+    "DecisionCodeEdge",
+    "DecisionCreate",
+    "DecisionEvidenceResponse",
+    "DecisionGraphEdge",
+    "DecisionGraphNode",
+    "DecisionGraphResponse",
+    "DecisionLineageEntry",
+    "DecisionRecordResponse",
+    "DecisionStatusUpdate",
+    "DirectRiskEntry",
+    "EgoGraphResponse",
+    "ExecutionFlowEntry",
+    "ExecutionFlowsResponse",
+    "GitMetadataResponse",
+    "GitSummaryResponse",
+    "GraphEdgeResponse",
+    "GraphExportResponse",
+    "GraphMetricsResponse",
+    "GraphNodeResponse",
+    "HealthResponse",
+    "HotFilesGraphResponse",
+    "HotFilesNodeResponse",
+    "HotspotResponse",
+    "JobResponse",
+    "KnowledgeMapOwner",
+    "KnowledgeMapResponse",
+    "KnowledgeMapSilo",
+    "KnowledgeMapTarget",
+    "ModuleEdgeResponse",
+    "ModuleGraphResponse",
+    "ModuleHealthDetail",
+    "ModuleHealthOwner",
+    "ModuleHealthSummary",
+    "ModuleNodeResponse",
+    "NeighboringCommunity",
+    "NodeSearchResult",
+    "OwnerCoAuthor",
+    "OwnerFileEntry",
+    "OwnerListEntry",
+    "OwnerModuleRollup",
+    "OwnerProfileResponse",
+    "OwnershipEntry",
+    "PageResponse",
+    "PageVersionResponse",
+    "Paginated",
+    "RepoCreate",
+    "RepoResponse",
+    "RepoStatsResponse",
+    "RepoUpdate",
+    "ReviewerEntry",
+    "ReviewerSuggestion",
+    "ReviewerSuggestionsResponse",
+    "SearchRequest",
+    "SearchResultResponse",
+    "SecurityFindingResponse",
+    "SetActiveProviderRequest",
+    "SetApiKeyRequest",
+    "SymbolImportanceComponents",
+    "SymbolNodeSummary",
+    "SymbolResponse",
+    "TransitiveEntry",
+    "WebhookResponse",
+    "WorkspaceCoChangeEntry",
+    "WorkspaceCoChangesResponse",
+    "WorkspaceContractEntry",
+    "WorkspaceContractLinkEntry",
+    "WorkspaceContractSummary",
+    "WorkspaceContractsResponse",
+    "WorkspaceCrossRepoSummary",
+    "WorkspaceGraphEdge",
+    "WorkspaceGraphNode",
+    "WorkspaceGraphResponse",
+    "WorkspaceRepoEntry",
+    "WorkspaceResponse",
+    "WorkspaceSyncResponse",
+    "WorkspaceSyncResult",
+]
